Revert to executable-based backend with STDOUT communication
