+++ gui/pairinteraction/worker.py
@@ -48,7 +48,7 @@ class Worker(QtCore.QThread):
             self.dataqueue_potential.queue.clear()
 
     def run(self):
-        finishedgracefully = True
+        finishedgracefully = False
 
         self.message = ""
 
@@ -75,12 +75,12 @@ class Worker(QtCore.QThread):
         status_progress = ""
         status_dimension = ""
 
-        for line in self.stdout:
+        for line in iter(self.stdout.readline, b""):
             if self.exiting or not line:
                 break
 
-            elif line[:5] == u">>TYP":
-                type = int(line[5:12])
+            elif line[:5] == b">>TYP":
+                type = int(line[5:12].decode('utf-8'))
                 status_type = ["Field map of first atom: ",
                                "Field map of second atom: ", "Pair potential: ", "Field maps: "][type]
                 status_progress = "construct matrices"
@@ -90,13 +90,13 @@ class Worker(QtCore.QThread):
                 elif type == 0 or type == 1:
                     self.samebasis = False
 
-            elif line[:5] == u">>BAS":
-                basissize = int(line[5:12])
+            elif line[:5] == b">>BAS":
+                basissize = int(line[5:12].decode('utf-8'))
                 status_progress = "construct matrices using {} basis vectors".format(
                     basissize)
 
-            elif line[:5] == u">>STA":
-                filename = line[6:-1].strip()
+            elif line[:5] == b">>STA":
+                filename = line[6:-1].decode('utf-8').strip()
                 if type == 0 or type == 3:
                     self.basisfile_field1 = filename
                 elif type == 1:
@@ -104,25 +104,25 @@ class Worker(QtCore.QThread):
                 elif type == 2:
                     self.basisfile_potential = filename
 
-            elif line[:5] == u">>TOT":
-                total = int(line[5:12])
+            elif line[:5] == b">>TOT":
+                total = int(line[5:12].decode('utf-8'))
                 current = 0
 
-            elif line[:5] == u">>DIM":
+            elif line[:5] == b">>DIM":
                 dim = int(line[5:12])
                 status_progress = "diagonalize {} x {} matrix, {} of {} matrices processed".format(
                     dim, dim, current, total)
 
-            elif line[:5] == u">>OUT":
+            elif line[:5] == b">>OUT":
                 current += 1
                 status_progress = "diagonalize {} x {} matrix, {} of {} matrices processed".format(
                     dim, dim, current, total)
 
-                filenumber = int(line[5:12])
-                filestep = int(line[12:19])
-                blocks = int(line[19:26])
-                blocknumber = int(line[26:33])
-                filename = line[34:-1].strip()
+                filenumber = int(line[5:12].decode('utf-8'))
+                filestep = int(line[12:19].decode('utf-8'))
+                blocks = int(line[19:26].decode('utf-8'))
+                blocknumber = int(line[26:33].decode('utf-8'))
+                filename = line[34:-1].decode('utf-8').strip()
 
                 if type == 0 or type == 3:
                     self.dataqueue_field1.put(
@@ -134,11 +134,15 @@ class Worker(QtCore.QThread):
                     self.dataqueue_potential.put(
                         [filestep, blocks, blocknumber, filename])
 
-            elif line[:5] == u">>ERR":
-                self.criticalsignal.emit(line[5:].strip())
+            elif line[:5] == b">>ERR":
+                self.criticalsignal.emit(line[5:].decode('utf-8').strip())
+
+            elif line[:5] == b">>END":
+                finishedgracefully = True
+                break
 
             else:
-                print(line)
+                print(line.decode('utf-8'), end="")
 
             self.message = status_type + status_progress
